backend/requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
sqlalchemy
pgvector
psycopg[binary]
pypdf
streamlit
phidata
google-generativeai
fastapi 
uvicorn 
pydantic 
SQLAlchemy 
alembic
crewai
celery 
redis
fastapi[standard]
crewai_tools